Don't crash when we call __qualname__ on torch._C.ScriptFunction
We've root caused this to correctly throwing attribute error on ScriptFunction
when missing attributes are caused. This PR will fix crashes that are showing
up. I'm going to stack a second PR to fix torch._c.ScriptFunction just being a
very badly behaving python object (which should also fix this

ghstack-source-id: 86bc3b5c062f67a467f34d9d4c393763062809bf
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147894