scriptfunction: Make sure we have valid __name__ and __qualname__
It's not fully clear why these are not being created, but you can definitely
reproduce this in code. __name__ is fun, since there appears to be no way to
explicitly set it on the pybind11 layer or c++ layer. I've set this in the
python wrapper code (which works correctly). But let me know if people feel
strongly and want us to go explicitly cast to python within the cpp functions
and set it there.

ghstack-source-id: c9241c3079d5bbce6341b26d05bcf612c019b6fd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/147906